+++ pkg/api/service.go
@@ -3,6 +3,8 @@ package api
 import (
 	"github.com/fiwippi/tanuki/pkg/api/admin"
 	"github.com/fiwippi/tanuki/pkg/api/auth"
+	"github.com/fiwippi/tanuki/pkg/api/manager"
+	"github.com/fiwippi/tanuki/pkg/api/mangadex"
 	"github.com/fiwippi/tanuki/pkg/api/series"
 	"github.com/fiwippi/tanuki/pkg/api/user"
 	"github.com/fiwippi/tanuki/pkg/server"
@@ -15,6 +17,6 @@ func NewService(s *server.Instance) {
 	auth.NewService(api)
 	series.NewService(api)
 	user.NewService(api)
-	//download.NewService(api)
-	//proxy.NewService(api)
+	mangadex.NewService(api)
+	manager.NewService(api)
 }